+++ spec/lib/tmuxinator/cli_spec.rb
@@ -182,7 +182,7 @@ describe Tmuxinator::Cli do
 
       before do
         # make sure that no project file exists initially
-        FileUtils.remove_file(path) if File.exists?(path)
+        FileUtils.remove_file(path) if File.exist?(path)
         expect(File).not_to exist(path)
 
         # now generate a project file
@@ -222,7 +222,7 @@ describe Tmuxinator::Cli do
 
       context "existing project doesn't exist" do
         before do
-          expect(File).to receive_messages(exists?: false)
+          expect(File).to receive_messages(exist?: false)
         end
 
         it "creates a new tmuxinator project file" do
@@ -235,8 +235,8 @@ describe Tmuxinator::Cli do
         let(:root_path) { "#{ENV['HOME']}\/\.tmuxinator\/#{name}\.yml" }
 
         before do
-          allow(File).to receive(:exists?).with(anything).and_return(false)
-          expect(File).to receive(:exists?).with(root_path).and_return(true)
+          allow(File).to receive(:exist?).with(anything).and_return(false)
+          expect(File).to receive(:exist?).with(root_path).and_return(true)
         end
 
         it "just opens the file" do
@@ -253,7 +253,7 @@ describe Tmuxinator::Cli do
 
       context "existing project doesn't exist" do
         before do
-          allow(File).to receive(:exists?).at_least(:once) do
+          allow(File).to receive(:exist?).at_least(:once) do
             false
           end
         end
@@ -267,7 +267,7 @@ describe Tmuxinator::Cli do
       context "files exists" do
         let(:path) { Tmuxinator::Config::LOCAL_DEFAULT }
         before do
-          expect(File).to receive(:exists?).with(path) { true }
+          expect(File).to receive(:exist?).with(path) { true }
         end
 
         it "just opens the file" do
@@ -490,7 +490,7 @@ describe Tmuxinator::Cli do
     let(:path) { Tmuxinator::Config.default_project(name) }
 
     after(:each) do
-      FileUtils.remove_file(path) if File.exists?(path)
+      FileUtils.remove_file(path) if File.exist?(path)
     end
 
     context "when the project file does not already exist" do
@@ -538,7 +538,7 @@ describe Tmuxinator::Cli do
     end
 
     after(:each) do
-      FileUtils.remove_file(path) if File.exists?(path)
+      FileUtils.remove_file(path) if File.exist?(path)
     end
 
     it "should always generate a project file" do